Move more fallback related code into fallback
